refactor(canvas): drop the extra types and session restore
The bug is one PopScope callback treating rotation as back. Ignore pops
for 600ms after a metrics/size change; do not persist the open canvas,
wrap PopScope, or change the web manifest.
